+++ tests/Aws/Tests/Common/Client/AbstractClientTest.php
@@ -18,6 +18,7 @@ namespace Aws\Tests\Common\Client;
 
 use Aws\Common\Aws;
 use Aws\Common\Client\AbstractClient;
+use Aws\Common\Client\AwsClientInterface;
 use Aws\Common\Signature\SignatureV4;
 use Aws\Common\Signature\SignatureListener;
 use Aws\Common\Credentials\Credentials;
@@ -132,11 +133,20 @@ class AbstractClientTest extends \Guzzle\Tests\GuzzleTestCase
 
     public function testSetRegionUpdatesBaseUrlAndSignature()
     {
-        $client = $this->getServiceBuilder()->get('dynamodb');
+        /** @var $client AwsClientInterface */
+        $client = $this->getServiceBuilder()->get('dynamodb', true);
+
         $client->setRegion('us-west-1');
         $this->assertEquals('https://dynamodb.us-west-1.amazonaws.com', (string) $client->getBaseUrl());
         $this->assertEquals('https://dynamodb.us-west-1.amazonaws.com', $client->getConfig('base_url'));
+        $this->assertEquals('us-west-1', $client->getRegion());
         $this->assertEquals('us-west-1', $this->readAttribute($client->getSignature(), 'regionName'));
+
+        $client->setRegion('us-west-2');
+        $this->assertEquals('https://dynamodb.us-west-2.amazonaws.com', (string) $client->getBaseUrl());
+        $this->assertEquals('https://dynamodb.us-west-2.amazonaws.com', $client->getConfig('base_url'));
+        $this->assertEquals('us-west-2', $client->getRegion());
+        $this->assertEquals('us-west-2', $this->readAttribute($client->getSignature(), 'regionName'));
     }
 
     public function testAllowsMagicWaiters()